# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Bravo API
  version: 1.0.0
  description: OpenAPI documentation for the Bravo backend.
paths:
...
  /api/organization/:
    get:
      operationId: organization_list
      description: Lists all organizations. Only available to admin users.
      summary: List organizations
      tags:
      - organization
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
          description: ''
...
  /api/organization/me/:
    get:
      operationId: organization_me_retrieve
      description: Returns the organization associated with the authenticated user.
      summary: Get my organization
      tags:
      - organization
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
    patch:
      operationId: organization_me_partial_update
      description: Partially updates the organization associated with the authenticated
        user.
      summary: Update my organization
      tags:
      - organization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOrganization'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedOrganization'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOrganization'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
...
components:
  schemas:
...
    Organization:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 255
        legal_name:
          type: string
          maxLength: 255
        tax_id:
          type: string
          maxLength: 64
        billing_email:
          type: string
          format: email
          maxLength: 254
        billing_address:
          type: string
          maxLength: 255
        billing_city:
          type: string
          maxLength: 120
        billing_country:
          type: string
          maxLength: 2
        billing_postal_code:
          type: string
          maxLength: 20
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - billing_address
      - billing_city
      - billing_country
      - billing_email
      - billing_postal_code
      - created_at
      - legal_name
      - name
      - tax_id
      - updated_at
      - uuid
...
    PatchedOrganization:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 255
        legal_name:
          type: string
          maxLength: 255
        tax_id:
          type: string
          maxLength: 64
        billing_email:
          type: string
          format: email
          maxLength: 254
        billing_address:
          type: string
          maxLength: 255
        billing_city:
          type: string
          maxLength: 120
        billing_country:
          type: string
          maxLength: 2
        billing_postal_code:
          type: string
          maxLength: 20
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT